Migrate contract methods and update tests
