+++ sdks/open-format/src/core/sdk.ts
@@ -1,4 +1,3 @@
-import { BigNumberish } from 'ethers';
 import merge from 'lodash.merge';
 import {
   getProviderFromUrl,
@@ -65,8 +64,6 @@ export class OpenFormatSDK extends BaseContract {
       nft,
     });
 
-    this.options.contractAddress = tx.contractAddress;
-
     return tx;
   }
 
@@ -78,306 +75,6 @@ export class OpenFormatSDK extends BaseContract {
   getNFT(address: string) {
     invariant(this.signer, 'No signer set, cannot get NFT');
     return new OpenFormatNFT(address, this.provider, this.signer);
-  }
-
-  /**
-   * Mints an NFT with commission on a contract address
-   * @param {string} address - address of the contract
-   * @returns transaction
-   */
-  async mintWithCommission(address: string) {
-    invariant(this.signer, 'No signer set, aborting mint');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.mintWithCommission({
-      address,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Setup royalties to be paid to an address
-   * @param {Object} params
-   * @param {number} params.royaltyReceiverAddress - address of the receiver of the royalties
-   * @param {number} params.royaltyPercentage - the percentage between 0-1000 they will receive
-   * @returns
-   */
-  async setRoyalties(params: {
-    royaltyReceiverAddress: string;
-    royaltyPercentage: number;
-  }) {
-    invariant(this.signer, 'No signer set, aborting setting royalties');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.setRoyalties({
-      ...params,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Gets royalties based on sale price
-   * @param {Object} params
-   * @param {number} params.salePrice - sale price to get
-   * @returns
-   */
-  async getRoyalties({ salePrice }: { salePrice: number }) {
-    invariant(this.signer, 'No signer set, aborting getting royalties');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.getRoyalties({
-      salePrice,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Sets up Revenue Sharing
-   * @param {Object} params
-   * @param {string} params.contractAddress - address of the contract
-   * @param {{
-   *  address: string;
-   *  share: BigNumberish;
-   * }[]} params.collaborators - list of collaborators addresses
-   * @param {number} params.holderPercentage - the holders percentage
-   * @returns
-   */
-  async setupRevenueSharing(params: {
-    revShareExtensionAddress: string;
-    collaborators: {
-      address: string;
-      share: BigNumberish;
-    }[];
-    holderPercentage: BigNumberish;
-  }) {
-    invariant(this.signer, 'No signer set, aborting revenue sharing setup');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.setupRevenueSharing({
-      ...params,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Allocation of shares
-   * @param {{
-   *  address: string;
-   *  share: BigNumberish;
-   * }[]} params - array of account addresses and shares
-   * @returns
-   */
-  async allocateRevenueShares(
-    params: {
-      address: string;
-      share: BigNumberish;
-    }[]
-  ) {
-    invariant(this.signer, 'No signer set, aborting revenue sharing setup');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.allocateRevenueShares({
-      allocation: params,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Get the balance of a collaborator
-   * @param {string} address - Address of the collaborator
-   * @returns BigNumber
-   */
-  async getCollaboratorBalance(address: string) {
-    invariant(this.signer, 'No signer set, aborting revenue sharing setup');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.getCollaboratorBalance({
-      address,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Withdrawl of collaborator funds
-   * @param {string} address - Address of the collaborator
-   * @returns
-   */
-  async withdrawCollaboratorFunds(address: string) {
-    invariant(
-      this.signer,
-      'No signer set, aborting withdrawl of collaborator funds'
-    );
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.withdrawCollaboratorFunds({
-      address,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Get the balance of a token
-   * @param {BigNumberish} token - Token number
-   * @returns BigNumber
-   */
-  async getTokenBalance(token: BigNumberish) {
-    invariant(this.signer, 'No signer set, aborting get token balance');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.getTokenBalance({
-      token,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Withdrawl of token funds
-   * @param {BigNumberish} token - Token number
-   * @returns
-   */
-  async withdrawTokenFunds(token: BigNumberish) {
-    invariant(this.signer, 'No signer set, aborting withdrawl of token funds');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.withdrawTokenFunds({
-      token,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Sets the percentage of the primary commission
-   * @param {{ percent: BigNumberish; }} percent - Percent
-   * @returns
-   */
-  async setPrimaryCommissionPercent(percent: BigNumberish) {
-    invariant(
-      this.signer,
-      'No signer set, aborting set primary commission percent'
-    );
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.setPrimaryCommissionPercent({
-      percent,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Sets the percentages of the secondary commission
-   * @param {BigNumberish} percent - Percent
-   * @returns
-   */
-  async setSecondaryCommissionPercent(percent: BigNumberish) {
-    invariant(
-      this.signer,
-      'No signer set, aborting set secondary commission percent'
-    );
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.setSecondaryCommissionPercent({
-      percent,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  async setTokenSalePrice({
-    tokenId,
-    price,
-  }: {
-    tokenId: BigNumberish;
-    price: BigNumberish;
-  }) {
-    invariant(this.signer, 'No signer set, aborting set token sale price');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.setTokenSalePrice({
-      tokenId,
-      price,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Buy
-   * @param {Object} params
-   * @param {number} params.tokenId - id of the token
-   * @returns
-   */
-  async buy(tokenId: BigNumberish) {
-    invariant(this.signer, 'No signer set, aborting buy');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.buy({
-      tokenId,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
-  }
-
-  /**
-   * Buy with commission
-   * @param {Object} params
-   * @param {number} params.tokenId - id of the token
-   * @param {number} params.address - Address
-   * @returns
-   */
-  async buyWithCommission({
-    tokenId,
-    address,
-  }: {
-    tokenId: BigNumberish;
-    address: string;
-  }) {
-    invariant(this.signer, 'No signer set, aborting buy with commission');
-    invariant(this.options.contractAddress, 'No contract address set');
-
-    await this.checkNetworksMatch();
-
-    return contract.buyWithCommission({
-      tokenId,
-      address,
-      contractAddress: this.options.contractAddress,
-      signer: this.signer,
-    });
   }
 
   rawRequest = subgraph.rawRequest;
